Moving rollup related function out of constants to rollup-helper
